+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Obsolete dropdownItems() and dropdownItemsExt() unused functions
     - menuItems() private function refactoring to better handle the reactivity
     - Make sure rendered modal data context holds the parent AC object
+    - Define new 'signinSubmitFn()' and 'signinTitle' acUserLogin configuration options (used by pwix:accounts-zimbra)
 
 ### 2.2.0